requirements.txt
Flask==0.11.1
GoogleAppEngineCloudStorageClient==1.9.22.1
gunicorn==19.6.0
six==1.10.0
google-api-python-client==1.5.5
google-auth
requests[security]
